# packages/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Creator {
  id: string;
  name: string;
  location: string;
  niche: string;
  followers: number;
  currency: string;
  /** Stable key for the bundled creator avatar (e.g. "creator-1") */
  imageKey: string;
}

export interface TrendBrief {
  id: string;
  title: string;
  context: string;
  /**
   * @minimum 0
   * @maximum 100
   */
  viralPotential: number;
  description: string;
  imageKey: string;
}

export interface TrendBriefList {
  briefs: TrendBrief[];
}

export type BrandDealStatus =
  (typeof BrandDealStatus)[keyof typeof BrandDealStatus];

export const BrandDealStatus = {
  Signed: "Signed",
  Negotiating: "Negotiating",
  Paid: "Paid",
} as const;

export interface BrandDeal {
  id: string;
  brand: string;
  status: BrandDealStatus;
  amount: number;
}

export type AgentStatus = (typeof AgentStatus)[keyof typeof AgentStatus];

export const AgentStatus = {
  pending: "pending",
  active: "active",
  done: "done",
} as const;

export type VideoStatus = (typeof VideoStatus)[keyof typeof VideoStatus];

export const VideoStatus = {
  Ideating: "Ideating",
  Editing: "Editing",
  Ready: "Ready",
} as const;

export type VideoAgents = {
  Ideator: AgentStatus;
  Director: AgentStatus;
  Editor: AgentStatus;
  Monetizer: AgentStatus;
};

export interface Video {
  id: string;
  title: string;
  status: VideoStatus;
  /**
   * @minimum 0
   * @maximum 100
   */
  viralScore?: number | null;
  reasoning: string;
  thumbnailKey: string;
  script: string;
  agents: VideoAgents;
}

export interface VideoList {
  videos: Video[];
}

export interface SwarmRunStart {
  runId: string;
}

export type SwarmRunStatus =
  (typeof SwarmRunStatus)[keyof typeof SwarmRunStatus];

export const SwarmRunStatus = {
  queued: "queued",
  running: "running",
  done: "done",
  failed: "failed",
} as const;

export type AgentName = (typeof AgentName)[keyof typeof AgentName];

export const AgentName = {
  ideator: "ideator",
  director: "director",
  editor: "editor",
  monetizer: "monetizer",
} as const;

export interface SwarmAgentRun {
  agent: AgentName;
  status: SwarmRunStatus;
  summary?: string | null;
  error?: string | null;
  startedAt?: string | null;
  finishedAt?: string | null;
}

export interface SwarmRunSummary {
  id: string;
  /** Always 'swarm' for parent runs */
  agent: string;
  status: SwarmRunStatus;
  summary?: string | null;
  error?: string | null;
  startedAt?: string | null;
  finishedAt?: string | null;
  createdAt: string;
}

export interface SwarmRunList {
  runs: SwarmRunSummary[];
}

export type PublicationPlatform =
  (typeof PublicationPlatform)[keyof typeof PublicationPlatform];

export const PublicationPlatform = {
  tiktok: "tiktok",
  reels: "reels",
  shorts: "shorts",
} as const;

export type PublicationStatus =
  (typeof PublicationStatus)[keyof typeof PublicationStatus];

export const PublicationStatus = {
  queued: "queued",
  published: "published",
  failed: "failed",
  blocked: "blocked",
} as const;

export type ShieldVerdict = (typeof ShieldVerdict)[keyof typeof ShieldVerdict];

export const ShieldVerdict = {
  pass: "pass",
  rewritten: "rewritten",
  blocked: "blocked",
} as const;

export interface PublicationMetrics {
  /** @minimum 0 */
  views: number;
  /** @minimum 0 */
  likes: number;
  /** @minimum 0 */
  comments: number;
  /** @minimum 0 */
  shares: number;
}

export interface ConsentState {
  aiDisclosureConsentedAt: string | null;
  adultConfirmedAt: string | null;
}

export interface ConsentInput {
  aiDisclosureConsented: boolean;
  adultConfirmed: boolean;
}

export interface ScheduleState {
  enabled: boolean;
  /**
   * @minimum 0
   * @maximum 23
   */
  hour: number | null;
  /** @maxLength 64 */
  tz: string | null;
  lastRunAt: string | null;
}

export interface ScheduleInput {
  enabled: boolean;
  /**
   * @minimum 0
   * @maximum 23
   */
  hour: number | null;
  /** @maxLength 64 */
  tz: string | null;
}

export interface Publication {
  id: string;
  videoId: string;
  platform: PublicationPlatform;
  status: PublicationStatus;
  /** @maxLength 255 */
  platformPostId?: string | null;
  /** @maxLength 2048 */
  mockUrl?: string | null;
  scheduledFor?: string | null;
  publishedAt?: string | null;
  /** @maxLength 2000 */
  error?: string | null;
  shieldVerdict?: ShieldVerdict | null;
  metrics?: PublicationMetrics | null;
  metricsFetchedAt?: string | null;
  createdAt: string;
}

export interface PublicationList {
  publications: Publication[];
}

export interface CreatePublicationInput {
  platform: PublicationPlatform;
  status: PublicationStatus;
  shieldVerdict: ShieldVerdict;
  /** @maxLength 255 */
  platformPostId?: string | null;
  /** @maxLength 2048 */
  mockUrl?: string | null;
  scheduledFor?: string | null;
  publishedAt?: string | null;
  /** @maxLength 2000 */
  error?: string | null;
}

export interface SwarmRunDetail {
  id: string;
  status: SwarmRunStatus;
  summary?: string | null;
  error?: string | null;
  startedAt?: string | null;
  finishedAt?: string | null;
  createdAt: string;
  agents: SwarmAgentRun[];
}

export interface EarningsSummary {
  currentMonth: number;
  currency: string;
  growth: string;
  deals: BrandDeal[];
  history: number[];
}

export type ExportMyData200 = { [key: string]: unknown };

export type DeleteMyData200 = {
  deleted: boolean;
};