# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agents/runs/{id}:
    get:
      operationId: getSwarmRun
      tags: [agents]
      summary: Detailed view of one swarm run with per-agent status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Swarm run detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwarmRunDetail"
        "404":
          description: Run not found
  /videos/{id}/publications:
    post:
      operationId: recordPublication
      tags: [publications]
      summary: Record one platform publish outcome for a video
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePublicationInput"
      responses:
        "201":
          description: Publication recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Publication"
        "400":
          description: Invalid input
        "401":
          description: Unknown user
        "404":
          description: Video not found
    get:
      operationId: listVideoPublications
      tags: [publications]
      summary: List publications for a video
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Publications for the video
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicationList"
        "401":
          description: Unknown user
...
  /me/consent:
    get:
      operationId: getConsent
      tags: [me]
      summary: Current AI-disclosure + adult-confirmation consent state
      responses:
        "200":
          description: Consent state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsentState"
        "401":
          description: Unknown user
    post:
      operationId: upsertConsent
      tags: [me]
      summary: Grant or withdraw AI-disclosure / adult consent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsentInput"
      responses:
        "200":
          description: Updated consent state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsentState"
        "400":
          description: Invalid input
        "401":
          description: Unknown user
  /me/schedule:
    get:
      operationId: getSchedule
      tags: [me]
      summary: Nightly swarm schedule preferences
      responses:
        "200":
          description: Schedule state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleState"
        "401":
          description: Unknown user
    post:
      operationId: upsertSchedule
      tags: [me]
      summary: Set nightly swarm schedule preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleInput"
      responses:
        "200":
          description: Updated schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleInput"
        "400":
          description: Invalid input
        "401":
          description: Unknown user
        "403":
          description: Consent required to enable
  /me/data-export:
    post:
      operationId: exportMyData
      tags: [me]
      summary: GDPR/CCPA — every row owned by this creator as JSON
      responses:
        "200":
          description: Data export
          content:
            application/json:
              schema:
                type: object
        "401":
          description: Unknown user
  /me/data-delete:
    post:
      operationId: deleteMyData
      tags: [me]
      summary: GDPR/CCPA — hard-delete this creator and all owned rows
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
                required: [deleted]
        "401":
          description: Unknown user
  /publications/recent:
    get:
      operationId: listRecentPublications
      tags: [publications]
      summary: Recent publications across all videos for the signed-in creator
      responses:
        "200":
          description: Recent publications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicationList"
        "401":
          description: Unknown user
components:
  schemas:
...
    SwarmRunStatus:
      type: string
      enum: [queued, running, done, failed]
    AgentName:
      type: string
      enum: [ideator, director, editor, monetizer]
    SwarmAgentRun:
      type: object
      properties:
        agent:
          $ref: "#/components/schemas/AgentName"
        status:
          $ref: "#/components/schemas/SwarmRunStatus"
        summary:
          type: string
          nullable: true
        error:
          type: string
          nullable: true
        startedAt:
          type: string
          nullable: true
        finishedAt:
          type: string
          nullable: true
      required: [agent, status]
...
    PublicationPlatform:
      type: string
      enum: [tiktok, reels, shorts]
    PublicationStatus:
      type: string
      enum: [queued, published, failed, blocked]
    ShieldVerdict:
      type: string
      enum: [pass, rewritten, blocked]
    PublicationMetrics:
      type: object
      properties:
        views:
          type: integer
          minimum: 0
        likes:
          type: integer
          minimum: 0
        comments:
          type: integer
          minimum: 0
        shares:
          type: integer
          minimum: 0
      required: [views, likes, comments, shares]
    ConsentState:
      type: object
      properties:
        aiDisclosureConsentedAt:
          type: string
          format: date-time
          nullable: true
        adultConfirmedAt:
          type: string
          format: date-time
          nullable: true
      required: [aiDisclosureConsentedAt, adultConfirmedAt]
    ConsentInput:
      type: object
      properties:
        aiDisclosureConsented:
          type: boolean
        adultConfirmed:
          type: boolean
      required: [aiDisclosureConsented, adultConfirmed]
    ScheduleState:
      type: object
      properties:
        enabled:
          type: boolean
        hour:
          type: integer
          minimum: 0
          maximum: 23
          nullable: true
        tz:
          type: string
          maxLength: 64
          nullable: true
        lastRunAt:
          type: string
          format: date-time
          nullable: true
      required: [enabled, hour, tz, lastRunAt]
    ScheduleInput:
      type: object
      properties:
        enabled:
          type: boolean
        hour:
          type: integer
          minimum: 0
          maximum: 23
          nullable: true
        tz:
          type: string
          maxLength: 64
          nullable: true
      required: [enabled, hour, tz]
    Publication:
      type: object
      properties:
        id:
          type: string
        videoId:
          type: string
        platform:
          $ref: "#/components/schemas/PublicationPlatform"
        status:
          $ref: "#/components/schemas/PublicationStatus"
        platformPostId:
          type: string
          maxLength: 255
          nullable: true
        mockUrl:
          type: string
          maxLength: 2048
          nullable: true
        scheduledFor:
          type: string
          format: date-time
          nullable: true
        publishedAt:
          type: string
          format: date-time
          nullable: true
        error:
          type: string
          maxLength: 2000
          nullable: true
        shieldVerdict:
          allOf:
            - $ref: "#/components/schemas/ShieldVerdict"
          nullable: true
        metrics:
          allOf:
            - $ref: "#/components/schemas/PublicationMetrics"
          nullable: true
        metricsFetchedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, videoId, platform, status, createdAt]
    PublicationList:
      type: object
      properties:
        publications:
          type: array
          items:
            $ref: "#/components/schemas/Publication"
      required: [publications]
    CreatePublicationInput:
      type: object
      properties:
        platform:
          $ref: "#/components/schemas/PublicationPlatform"
        status:
          $ref: "#/components/schemas/PublicationStatus"
        shieldVerdict:
          $ref: "#/components/schemas/ShieldVerdict"
        platformPostId:
          type: string
          maxLength: 255
          nullable: true
        mockUrl:
          type: string
          maxLength: 2048
          nullable: true
        scheduledFor:
          type: string
          format: date-time
          nullable: true
        publishedAt:
          type: string
          format: date-time
          nullable: true
        error:
          type: string
          maxLength: 2000
          nullable: true
      required: [platform, status, shieldVerdict]
    SwarmRunDetail:
      type: object
      properties:
        id:
          type: string
        status:
          $ref: "#/components/schemas/SwarmRunStatus"
        summary:
          type: string
          nullable: true
        error:
          type: string
          nullable: true
        startedAt:
          type: string
          nullable: true
        finishedAt:
          type: string
          nullable: true
        createdAt:
          type: string
        agents:
          type: array
          items:
            $ref: "#/components/schemas/SwarmAgentRun"
      required: [id, status, createdAt, agents]
...