# packages/api-zod/src/generated/types/createPublicationInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PublicationPlatform } from "./publicationPlatform";
import type { PublicationStatus } from "./publicationStatus";
import type { ShieldVerdict } from "./shieldVerdict";

export interface CreatePublicationInput {
  platform: PublicationPlatform;
  status: PublicationStatus;
  shieldVerdict: ShieldVerdict;
  /** @maxLength 255 */
  platformPostId?: string | null;
  /** @maxLength 2048 */
  mockUrl?: string | null;
  scheduledFor?: Date | null;
  publishedAt?: Date | null;
  /** @maxLength 2000 */
  error?: string | null;
}