# packages/api-zod/src/generated/types/publicationMetrics.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PublicationMetrics {
  /** @minimum 0 */
  views: number;
  /** @minimum 0 */
  likes: number;
  /** @minimum 0 */
  comments: number;
  /** @minimum 0 */
  shares: number;
}